More abstract code, moved abstract part to math Instances can now be controlled in Scene, giving an easier way to customize the instances to create

+++ math/OctreeNodeTriangleHandler.cpp
@@ -1,6 +1,6 @@
 #include "math.hpp"
 
-OctreeNodeTriangleHandler::OctreeNodeTriangleHandler(Geometry * chunk, int * count){
+OctreeNodeTriangleHandler::OctreeNodeTriangleHandler(Geometry * chunk, long * count){
     this->chunk = chunk;
     this->count = count;
 }
